Added check for id

+++ delete.php
@@ -3,7 +3,11 @@ require_once('database.php');
 require_once('head.php');
 $dbh = new databaseaccess;
 $id = $_POST["id"];
-$dbh->delete($id);
-echo "Post Deleted<br>";
+if(NULL === $id){
+	echo "Please Login";
+}else{
+	$dbh->delete($id);
+	echo "Post Deleted";
+}
 require_once('foot.php');
 ?>
